+++ .oxlintrc.json
@@ -66,7 +66,7 @@
         "**/auth/config.ts",
         "**/functions/middleware.ts",
         "**/workflow/scoped-workflow.ts",
-        "**/workflow/reconcile-all.ts",
+        "**/cron/reconcile-all.ts",
         "**/openrouter/callback.ts"
       ],
       "rules": {
